+++ frontend/wailsjs/go/app/App.js
@@ -1270,8 +1270,8 @@ export function TestOAuthConnection(arg1) {
   return window['go']['app']['App']['TestOAuthConnection'](arg1);
 }
 
-export function TestSMTPConnection(arg1, arg2, arg3, arg4, arg5) {
-  return window['go']['app']['App']['TestSMTPConnection'](arg1, arg2, arg3, arg4, arg5);
+export function TestSMTPConnection(arg1, arg2, arg3, arg4, arg5, arg6) {
+  return window['go']['app']['App']['TestSMTPConnection'](arg1, arg2, arg3, arg4, arg5, arg6);
 }
 
 export function Trash(arg1) {